pyproject.toml
[project]
name = "md-reheader"
version = "0.1.0"
description = "Fine-tuned Qwen3-0.6B for restoring heading hierarchy in markdown documents"
requires-python = ">=3.13"
license = "Apache-2.0"
dependencies = [
    "torch>=2.11",
    "transformers>=5.4",
    "datasets>=4.8",
    "wandb>=0.25",
    "hydra-core>=1.3.2",
    "evaluate>=0.4.6",
    "huggingface-hub>=1.8",
    "pydantic>=2.12",
    "numpy>=2.4",
    "accelerate>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "ruff>=0.15",
    "pre-commit>=4.5",
    "nbstripout>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/md_reheader"]

[tool.ruff]
src = ["src"]
line-length = 99
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["md_reheader"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
